Async prisma handler.
